requirements.txt
pandas==2.2.3
numpy==2.2.2
matplotlib==3.10.0
seaborn==0.13.2
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytest==7.4.2